WId is HWND* under windows. So for 32bit systems, cast away this problem. For 64bit problems this will still give us some headaches...
svn path=/trunk/KDE/kdebase/workspace/; revision=1123450

+++ libs/kworkspace/kactivityconsumer.cpp
@@ -67,12 +67,12 @@ QStringList KActivityConsumer::availableActivities() const
 
 void KActivityConsumer::registerResourceWindow(WId wid, const KUrl & uri)
 {
-    d->manager()->RegisterResourceWindow(wid, uri.url());
+    d->manager()->RegisterResourceWindow((uint)wid, uri.url());
 }
 
 void KActivityConsumer::unregisterResourceWindow(WId wid, const KUrl & uri)
 {
-    d->manager()->UnregisterResourceWindow(wid, uri.url());
+    d->manager()->UnregisterResourceWindow((uint)wid, uri.url());
 }
 
 QStringList KActivityConsumer::activitiesForResource(const KUrl & uri)
